Add files via upload
Fixing a bug in which duplicate links would cause a page encoding error. The replace method would replace both at the same time and then when it came time in the loops to do the second replace, it didnt exist. I have switched it to work based on a set. It seems to have resolved the issue.